get_data and get_accumulated_data function tests
